+++ src/MinifyCssJsBundle/Templating/Helper/HeadScript.php
@@ -69,7 +69,7 @@ class HeadScript extends \Pimcore\Templating\Helper\HeadScript {
                 $timestamps[] = filemtime(PIMCORE_WEB_ROOT.$src);
             }
             $file = PIMCORE_WEB_ROOT.$this->tmpPath.'/fc_'.md5(json_encode($combine)).".js";
-            if (is_file($file) && filemtime($file) < max($timestamps)) {
+            if (is_file($file) && count($timestamps) > 0 && filemtime($file) < max($timestamps)) {
                 unlink($file);
             }
         } else {
